+++ KelvinRepr/IndexerKelvinRepr1.h
@@ -7,6 +7,6 @@ class IndexerKelvinRepr1 : public IIndexer<int>
 {
 public:
 	IndexerKelvinRepr1();
-	int ToFlat(std::initializer_list<int> ijk) const override;
+	size_t ToFlat(std::initializer_list<int> ijk) const override;
 	int numberOfIndices() const override { return 2; }
 };
