+++ KelvinRepr/IndexerKelvinRepr2.h
@@ -7,7 +7,7 @@ class IndexerKelvinRepr2 : public IIndexer<int>
 {
 public:
 	IndexerKelvinRepr2();
-	int ToFlat(std::initializer_list<int> ijk) const override;
+	size_t ToFlat(std::initializer_list<int> ijk) const override;
 	int numberOfIndices() const override { return 2; }
 };
